+++ prover/src/tests/mod.rs
@@ -16,3 +16,5 @@ pub mod lt_bus_tests;
 pub mod lt_tests;
 #[cfg(test)]
 pub mod prove_elfs_tests;
+#[cfg(test)]
+pub mod trace_builder_tests;
